Add solution for exercise 4.21

+++ part4/blog-list/controllers/blogLists.js
@@ -7,14 +7,6 @@ const jwt = require("jsonwebtoken")
 
 let exportedList = []
 
-// const getTokenFrom = (req) => {
-//   const authorization = req.get("authorization")
-//   if (authorization && authorization.startsWith("Bearer ")) {
-//     return authorization.replace("Bearer ", "")
-//   }
-//   return null
-// }
-
 blogListsRouter.get("/", async (req, res) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
   res.json(blogs)
@@ -46,7 +38,32 @@ blogListsRouter.post("/", async (req, res) => {
 
 blogListsRouter.delete("/:id", async (req, res) => {
   const id = req.params.id
+  const { token } = req
+
+  if (!token) {
+    return res.status(401).json({ error: "token missing" })
+  }
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" })
+  }
+
+  const blog = await Blog.findById(id)
+
+  if (!blog) {
+    return res.status(404).json({ error: "blog not found" })
+  }
+
+  if (blog.user.toString() !== decodedToken.id) {
+    return res.status(401).json({ error: "Unauthorized" })
+  }
   await Blog.findByIdAndRemove(id)
+
+  const user = await User.findById(decodedToken.id)
+  user.blogs = user.blogs.filter((blog) => blog.toString() !== id)
+  await user.save()
+
   return res.status(204).end()
 })
